Moved classes that are considered high-level to the high-level package. Made new class inside CSS/Highlevel/Implementation that handles CSS data with static methods
