+++ Silhouette/src/CSS/HighLevel/Animation.java
@@ -1,4 +1,6 @@
-package CSS.LowLevel.Style;
+package CSS.HighLevel;
+
+import CSS.LowLevel.RuleSet;
 
 public class Animation extends RuleSet {
     //public void
