+++ Silhouette/src/CSS/HighLevel/Values/Gradient.java
@@ -1,0 +1,27 @@
+package CSS.HighLevel.Values;
+
+import CSS.HighLevel.Implementation.CSSManager;
+
+import java.util.Arrays;
+import java.util.Iterator;
+
+public class Gradient {
+    /**
+     * Returns a linear or radial gradient as String
+     * E.g Color.gradient("linear", "180degs", Color.RGB(), "yellow")
+     * @param type "linear" or "radial"
+     * @param values
+     * @return String
+     */
+    public static String linearGradient(String type, String... values){
+        String str = CSSManager.combineValuesWithSeparator(", ", values);
+
+        return "linear-gradient(" + str + ")";
+    }
+
+    public static String radialGradient(String type, String... values){
+        String str = CSSManager.combineValuesWithSeparator(", ", values);
+
+        return "radial-gradient(" + str + ")";
+    }
+}
